// frontend/src/backend.d.ts
export interface PayoutRecord {
    id: string;
    payoutMethod: string;
    destinationAccount: string;
    artistId: string;
    currency: string;
    notes?: string;
    associatedMediaId?: string;
    timestamp: bigint;
    amount: bigint;
}